I just hacked in extremely hacky Python3 C-API compatibility.
On the upside, it seems that Pymod is now Python3 compatible!

I provided an alternate definition of the Nim proc `extendWithPyModinitFunc` that is used when the conditional symbol `python3` is defined in the Nim compiler by "pmgen.py".

This conditional symbol `python3` is defined by "pmgen.py" if the new function `define_python3_maybe` determines that the Python version is >= 3.

A based my hacky code upon the ugly, gratuitously-complex example code at https://docs.python.org/3/howto/cporting.html